Switch to using blobs for fluentd
Rather than committing a bunch of gems which have to stay updated with a
script, instead just have a job that packages the gems from
fluentd.Gemfile and creates a blob as BOSH intended. This should allow
people to build the release locally and we'll just use Concourse to
update the blob for public consumption. Dependabot or Renovate can look
after the Gemfile hopefully for keeping the version of fluentd up to
date